[refactor] Move version_check out of taichi.lang.
