+++ python/taichi/lang/misc.py
@@ -1,16 +1,11 @@
 import atexit
-import datetime
 import functools
-import json
 import os
-import platform
 import shutil
 import tempfile
-import threading
 import time
 from contextlib import contextmanager
 from copy import deepcopy as _deepcopy
-from urllib import request
 
 from taichi._lib import core as _ti_core
 from taichi._lib.utils import locale_encode
@@ -24,7 +19,7 @@ from taichi.profiler.kernelmetrics import default_cupti_metrics
 from taichi.tools.util import set_gdb_trigger, warning
 from taichi.types.primitive_types import f32, f64, i32, i64
 
-from taichi import _logging, _snode
+from taichi import _logging, _snode, _version_check
 
 i = axes(0)
 j = axes(1)
@@ -415,89 +410,6 @@ def prepare_sandbox():
     return tmp_dir
 
 
-def check_version():
-    # Check Taichi version for the user.
-    major = _ti_core.get_version_major()
-    minor = _ti_core.get_version_minor()
-    patch = _ti_core.get_version_patch()
-    version = f'{major}.{minor}.{patch}'
-    payload = {'version': version, 'platform': '', 'python': ''}
-
-    system = platform.system()
-    if system == 'Linux':
-        payload['platform'] = 'manylinux1_x86_64'
-    elif system == 'Windows':
-        payload['platform'] = 'win_amd64'
-    elif system == 'Darwin':
-        if platform.release() < '19.0.0':
-            payload['platform'] = 'macosx_10_14_x86_64'
-        elif platform.machine() == 'x86_64':
-            payload['platform'] = 'macosx_10_15_x86_64'
-        else:
-            payload['platform'] = 'macosx_11_0_arm64'
-
-    python_version = platform.python_version()
-    if python_version.startswith('3.6.'):
-        payload['python'] = 'cp36'
-    elif python_version.startswith('3.7.'):
-        payload['python'] = 'cp37'
-    elif python_version.startswith('3.8.'):
-        payload['python'] = 'cp38'
-    elif python_version.startswith('3.9.'):
-        payload['python'] = 'cp39'
-
-    # We do not want request exceptions break users' usage of Taichi.
-    try:
-        payload = json.dumps(payload)
-        payload = payload.encode()
-        req = request.Request('https://metadata.taichi.graphics/check_version',
-                              method='POST')
-        req.add_header('Content-Type', 'application/json')
-        with request.urlopen(req, data=payload, timeout=5) as response:
-            response = json.loads(response.read().decode('utf-8'))
-            return response
-    except:
-        return None
-
-
-def try_check_version():
-    try:
-        os.makedirs(_ti_core.get_repo_dir(), exist_ok=True)
-        timestamp_path = os.path.join(_ti_core.get_repo_dir(), 'timestamp')
-        cur_date = datetime.date.today()
-        if os.path.exists(timestamp_path):
-            last_time = ''
-            with open(timestamp_path, 'r') as f:
-                last_time = f.readlines()[0].rstrip()
-            if cur_date.strftime('%Y-%m-%d') > last_time:
-                response = check_version()
-                if response is None:
-                    return
-                with open(timestamp_path, 'w') as f:
-                    f.write((cur_date +
-                             datetime.timedelta(days=7)).strftime('%Y-%m-%d'))
-                    f.write('\n')
-                    if response['status'] == 1:
-                        f.write(response['latest_version'])
-                    else:
-                        f.write('0.0.0')
-        else:
-            response = check_version()
-            if response is None:
-                return
-            with open(timestamp_path, 'w') as f:
-                f.write((cur_date +
-                         datetime.timedelta(days=7)).strftime('%Y-%m-%d'))
-                f.write('\n')
-                if response['status'] == 1:
-                    f.write(response['latest_version'])
-                else:
-                    f.write('0.0.0')
-    # Wildcard exception to catch potential file writing errors.
-    except:
-        pass
-
-
 def init(arch=None,
          default_fp=None,
          default_ip=None,
@@ -525,12 +437,7 @@ def init(arch=None,
             * ``packed`` (bool): Enables the packed memory layout. See https://docs.taichi.graphics/lang/articles/advanced/layout.
     """
     # Check version for users every 7 days if not disabled by users.
-    skip = os.environ.get("TI_SKIP_VERSION_CHECK")
-    if skip != 'ON':
-        # We don't join this thread because we do not wish to block users.
-        check_version_thread = threading.Thread(target=try_check_version,
-                                                daemon=True)
-        check_version_thread.start()
+    _version_check.start_version_check_thread()
 
     # Make a deepcopy in case these args reference to items from ti.cfg, which are
     # actually references. If no copy is made and the args are indeed references,
